refactor:Refactoring production code
Add the functionalites of handling fused data from both lidar and radar sensors for kalman filter tracking.

+++ src/measurement_package.h
@@ -5,14 +5,14 @@
 
 class MeasurementPackage {
 public:
-	long timestamp_;
+  long long timestamp_;
 
-	enum SensorType {
-		LASER, RADAR
-	} sensor_type_;
+  enum SensorType{
+    LASER,
+    RADAR
+  } sensor_type_;
 
-	Eigen::VectorXd raw_measurements_;
-
+  Eigen::VectorXd raw_measurements_;
 };
 
 #endif /* MEASUREMENT_PACKAGE_H_ */
